+++ tests/test_conv2d_growing_module.py
@@ -108,6 +108,26 @@ class TestConv2dMergeGrowingModule(TorchTestCase):
         m._input_volume = None
         self.assertEqual(m.input_volume, self.prev.output_volume)
 
+    def test_output_volume_with_reshaping(self):
+        """Test output_volume when the ouput is reshaped."""
+        m = self.merge
+        # Set post_merge_function
+        m.post_merge_function = torch.nn.AvgPool2d(2, 2)
+        output_size = compute_output_shape_conv(self.prev.input_size, self.prev)
+        output_size = (output_size[0] - 2) / 2 + 1
+        output_volume = m.in_channels * output_size * output_size
+        self.assertEqual(m.output_volume, output_volume)
+
+        # Set reshaping function
+        m.reshape_function = torch.nn.Flatten()
+        self.assertEqual(m.output_volume, output_volume)
+        m.reshape_function = torch.nn.AdaptiveAvgPool2d(output_size=1)
+        self.assertEqual(m.output_volume, m.in_channels)
+
+        # Reset
+        m.input_size = None
+        self.assertEqual(m.output_volume, self.prev.output_volume)
+
     def test_constructor_int_conversions(self):
         """Test that int input_size and next_kernel_size are converted to tuples."""
         # Test with int input_size